Added trends, improved code

+++ Python_Scripts/VT_API/use_cases/create_update_collections.py
@@ -1,8 +1,8 @@
 ################
 # Python Script to 
 # 1. create a collection (done)
-# 2. add in indicators from an advance search (partial)
-# 3. get the top 3 malware families, and top 3 threat categories (todo)
+# 2. add in indicators from an advance search (partial - to settle pagnation next)
+# 3. get the top 3 malware families, and top 3 threat categories (done)
 # 4. update results from #3 to a google sheets (todo)
 #
 # author: VirusTotal // dominicchua@google.com
@@ -20,8 +20,8 @@ import time
 ################
 # variables
 ################
-firstseen = "2024-04-15+"
-lastseen = "2024-04-30-"
+firstseen = "2024-04-21+"
+lastseen = "2024-04-27-"
 #firstseen = input("Enter First Seen Start Date (eg 2023-12-01+):")
 #lastseen = input("Enter First Seen End Date (eg 2023-12-31-):")
 LIMIT = '300'
@@ -35,8 +35,19 @@ COLLECTION_DESCRIPTION = "test description"
 #advance search
 ################
 def file(query): 
-    url = f'https://www.virustotal.com/api/v3/intelligence/search?query={urllib.parse.quote(query)}&limit={LIMIT}&descriptors_only=false'
+    url = f'https://www.virustotal.com/api/v3/intelligence/search?query={urllib.parse.quote(query)}&limit={LIMIT}&descriptors_only=true'
     headers = {'Accept': 'application/json', 'x-apikey': os.environ['VT_APIKEY']}
+#   while True:  # trying out 
+    res = requests.get(url, headers=headers)
+    res.raise_for_status()
+    data = res.json()
+    """
+#    meant for pagination 
+        if "links" in data and "next" in data["links"]:
+            url = data["links"]["next"]
+        else:
+            break  # Exit loop if no next page
+    """
     res = requests.get(url, headers=headers)
     res.raise_for_status()
     return res.json()
@@ -80,18 +91,68 @@ def get_collection(id):
     headers = {'Accept': 'application/json', 'x-apikey': os.environ['VT_APIKEY']}
     res = requests.get(url, headers=headers)
     res.raise_for_status()
-    print(res.text)
-    #return res.json()
+    #print(res.text)
+    return json.dumps(res.json())
+
+#######################
+#get top threats 
+# malware families, threat categories, c2 url, file types, collections
+#######################
+def print_top_trends(json_response):
+    # Parse JSON
+    data = json.loads(json_response)
+
+    # Extract top 3 malware families
+    malware_families = data['data']['attributes']['aggregations']['files']['malware_config_family_name'][:3]
+
+    # Extract top 3 threat categories
+    threat_categories = data['data']['attributes']['aggregations']['files']['popular_threat_category'][:3]
+
+    # Extract top 3 malware config C2 URLs
+    c2_urls = data['data']['attributes']['aggregations']['files']['malware_config_c2_url'][:3]
+
+    # Extract top 3 malware config C2 URLs
+    file_types = data['data']['attributes']['aggregations']['files']['file_types'][:3]
+
+    # Print top 3 popular malware families
+    print("Top 3 Malware Families:")
+    for mal in malware_families:
+        print(f"{mal['value']}: {mal['count']}")
+
+    # Print top 3 threat categories
+    print("\nTop 3 Threat Categories:")
+    for category in threat_categories:
+        print(f"{category['value']}: {category['count']}")
+
+    # Print top 3 C2 URLs
+    print("\nTop 3 Malware Config C2 URLs:")
+    for c2 in c2_urls:
+        print(f"{c2['value']}: {c2['count']}")
+
+    # Print top 3 C2 URLs
+    print("\nTop 3 Malware File Types:")
+    for file in file_types:
+        print(f"{file['value']}: {file['count']}")
 
 #######################
 #delete collection
 #######################
 def delete_collection(id):
-    url = f"https://www.virustotal.com/api/v3/collections/{id}"
-    headers = {'Accept': 'application/json', "content-type": "application/json", 'x-apikey': os.environ['VT_APIKEY']}
-    res = requests.delete(url, headers=headers)
-    res.raise_for_status()
-    print(res.text)
+    while True:
+        user_input = input("Do you want to delete the collection (Y/N):")
+        if user_input.lower() == 'y':
+            url = f"https://www.virustotal.com/api/v3/collections/{id}"
+            headers = {'Accept': 'application/json', "content-type": "application/json", 'x-apikey': os.environ['VT_APIKEY']}
+            res = requests.delete(url, headers=headers)
+            res.raise_for_status()
+            print(res.text)
+            print("\nDeleting Collection.")
+            break
+        elif user_input.lower() == 'n':
+            print("Collection not deleted. Link to Collection:", vt_col_link + id)
+            return
+        else:
+            print("Invalid input. Please enter 'y' or 'n'.")
 
 
 #######################
@@ -102,10 +163,16 @@ try:
     hashes = []  # Initialize an empty list to store hashes
     vt_col_link = "https://www.virustotal.com/gui/collection/" # Link to VT collection
 
+# Use only if descripters = true 
+#    if "data" in res:
+#        for item in res["data"]:
+#            if "attributes" in item and "sha256" in item["attributes"]:
+#                hashes.append(item["attributes"]["sha256"])
+
+# Use only when descriptors_only=false
     if "data" in res:
         for item in res["data"]:
-            if "attributes" in item and "sha256" in item["attributes"]:
-                hashes.append(item["attributes"]["sha256"])
+                hashes.append(item["id"])
 
     pprint(f"Total Number of Hashes: {len(hashes)}") #Print number of hashes
     collection_link = create_collection(COLLECTION_NAME, hashes)  # Get JSON response
@@ -114,8 +181,9 @@ try:
     print("Link to Collection:", vt_col_link + collection_id) # Print link to collection
     print("\nWaiting for 10 seconds...") #wait needed for commmonalities to be computed
     time.sleep(10)
-    get_collection(collection_id) # Get collection details
-    #delete_collection(collection_id) # cleanup collection during testing 
+    json_response = get_collection(collection_id) # Get collection details
+    print_top_trends(json_response) # Print top threats and IPs
+    delete_collection(collection_id) # cleanup collection during testing 
 
 ############
 #error handling
